T 28

+++ src/Controller/FriendshipController.php
@@ -3,8 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Friendship;
-use App\Entity\Profile;
-use App\Entity\User;
+use App\Repository\FriendshipRepository;
 use App\Repository\ProfileRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Exception;
@@ -19,26 +18,28 @@ class FriendshipController extends AbstractController
 {
     private EntityManagerInterface $entityManager;
     private ProfileRepository $profileRepository;
+    private FriendshipRepository $friendshipRepository;
 
-    public function __construct(EntityManagerInterface $entityManager, ProfileRepository $profileRepository)
+    public function __construct(EntityManagerInterface $entityManager, ProfileRepository $profileRepository, FriendshipRepository $friendshipRepository)
     {
         $this->entityManager = $entityManager;
         $this->profileRepository = $profileRepository;
+        $this->friendshipRepository = $friendshipRepository;
     }
 
-    #[Route('/request', name: 'send_friend_request', methods: ['POST'])]
-    public function sendFriendRequest(Request $request): JsonResponse
+    #[Route('/request/{profileId}', name: 'send_friend_request', methods: ['POST'])]
+    public function sendFriendRequest(int $profileId, Request $request): JsonResponse
     {
         try {
-            $requesterUsername = $request->headers->get('requester-username');
-            $receiverUsername = $request->headers->get('receiver-username');
+            $data = json_decode($request->getContent(), true);
+            $friendIdToRequest = $data["friendId"];
 
-            if (!$requesterUsername || !$receiverUsername) {
+            if (!$profileId || !$friendIdToRequest) {
                 return new JsonResponse('Invalid input.', Response::HTTP_BAD_REQUEST);
             }
 
-            $requesterProfileUser = $this->profileRepository->findOneBy(['username' => $requesterUsername]);
-            $receiverProfileUser = $this->profileRepository->findOneBy(['username' => $receiverUsername]);
+            $requesterProfileUser = $this->profileRepository->findOneBy(['id' => $profileId]);
+            $receiverProfileUser = $this->profileRepository->findOneBy(['id' => $friendIdToRequest]);
 
             if (!$requesterProfileUser || !$receiverProfileUser) {
                 return new JsonResponse('Requester or receiver not found.', Response::HTTP_NOT_FOUND);
@@ -67,7 +68,7 @@ class FriendshipController extends AbstractController
             $this->entityManager->flush();
 
             return new JsonResponse(['message' => 'Friend request sent.'], Response::HTTP_CREATED);
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             return new JsonResponse(['error' => 'An error occurred.'], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
@@ -86,7 +87,7 @@ class FriendshipController extends AbstractController
             $this->entityManager->flush();
 
             return new Response('Friend request accepted.', Response::HTTP_OK);
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             return new JsonResponse(['error' => 'An error occurred.'], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
@@ -105,80 +106,77 @@ class FriendshipController extends AbstractController
             $this->entityManager->flush();
 
             return new Response('Friend request rejected.', Response::HTTP_OK);
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             return new JsonResponse(['error' => 'An error occurred.'], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
 
     #[Route('/remove/{id}', name: 'remove_friend', methods: ['DELETE'])]
-    public function removeFriend(int $id): Response
+    public function removeFriend(int $id, Request $request): Response
     {
         try {
+            $data = json_decode($request->getContent(), true);
             $friendship = $this->entityManager->getRepository(Friendship::class)->find($id);
+            $requesterId = $data['requesterId'];
+            $receiverId = $data['receiverId'];
 
             if (!$friendship || 'accepted' !== $friendship->getStatus()) {
                 return new Response('Friendship not found or not accepted.', Response::HTTP_NOT_FOUND);
+            }
+
+            if (($friendship->getRequester()->getId() !== $requesterId && $friendship->getReceiver()->getId() !== $requesterId) ||
+                ($friendship->getRequester()->getId() !== $receiverId && $friendship->getReceiver()->getId() !== $receiverId)) {
+                return new Response('You are not authorized to remove this friendship.', Response::HTTP_FORBIDDEN);
             }
 
             $this->entityManager->remove($friendship);
             $this->entityManager->flush();
 
             return new Response('Friend removed.', Response::HTTP_OK);
-        } catch (\Exception $e) {
-            return new JsonResponse(['error' => 'An error occurred.'], Response::HTTP_INTERNAL_SERVER_ERROR);
+        } catch (Exception $e) {
+            return new JsonResponse(['error' => $e->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
 
-    #[Route('/list/{id}', name: 'list_friends', methods: ['GET'])]
-    public function listFriends(int $id): JsonResponse
+    #[Route('/list/{profileId}', name: 'list_friends', methods: ['GET'])]
+    public function listFriends(int $profileId, Request $request): JsonResponse
     {
         try {
-            $user = $this->entityManager->getRepository(User::class)->find($id);
+            $profile = $this->profileRepository->find($profileId);
 
-            if (!$user) {
-                return new JsonResponse(['error' => 'User not found.'], Response::HTTP_NOT_FOUND);
+            if (!$profile) {
+                return new JsonResponse(['error' => 'Profile not found.'], Response::HTTP_NOT_FOUND);
             }
 
-            $friendships = $this->entityManager->getRepository(Friendship::class)->findBy([
-                'requester' => $user,
-                'status' => 'accepted',
-            ]);
+            $page = max((int)$request->query->get('page', 1), 1);
+            $limit = max((int)$request->query->get('limit', 20), 1);
+            $offset = ($page - 1) * $limit;
+
+            $friendships = $this->friendshipRepository->findFriendshipUserProfileId($profileId, $limit, $offset);
 
             if (empty($friendships)) {
                 return new JsonResponse(['message' => 'No friends found.'], Response::HTTP_NOT_FOUND);
             }
 
-            $friends = array_map(function ($friendship) {
-                $receiverProfile = $friendship->getReceiver();
-                $receiverEmail = $receiverProfile->getUser() ? $receiverProfile->getUser()->getEmail() : null;
-
-                return [
-                    'id' => $receiverProfile->getId(),
-                    'email' => $receiverEmail,
-                    'username' => $receiverProfile->getUsername(),
-                ];
-            }, $friendships);
-
-            return new JsonResponse($friends, Response::HTTP_OK);
-
-        } catch (\Exception $e) {
-            return new JsonResponse(['error' => 'An error occurred.'], Response::HTTP_INTERNAL_SERVER_ERROR);
-
+            return new JsonResponse($friendships, Response::HTTP_OK);
+        } catch (Exception $e) {
+            return new JsonResponse(['error' => $e->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
 
-    #[Route('/list-requests/{userId}', name: 'list_friend_requests', methods: ['GET'])]
-    public function listFriendRequests(int $userId): JsonResponse
+    #[Route('/list-requests/{profileId}', name: 'list_friend_requests', methods: ['GET'])]
+    public function listFriendRequests(int $profileId): JsonResponse
     {
         try {
-            $user = $this->entityManager->getRepository(User::class)->find($userId);
+            $profile = $this->profileRepository->find($profileId);
 
-            if (!$user) {
-                return new JsonResponse('User not found.', Response::HTTP_NOT_FOUND);
+            if (!$profile) {
+                return new JsonResponse(['error' => 'Profile not found.'], Response::HTTP_NOT_FOUND);
             }
 
             $friendRequests = $this->entityManager->getRepository(Friendship::class)->findBy([
-                'receiver' => $user,
+                'receiver' => $profile,
+                'status' => 'pending',
             ]);
 
             if (0 === count($friendRequests)) {
@@ -202,7 +200,7 @@ class FriendshipController extends AbstractController
             }, $friendRequests);
 
             return new JsonResponse($requests, Response::HTTP_OK);
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             return new JsonResponse(['error' => 'An error occurred.'], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
